Feat: 전화번호 인증 관련 코드 분리
