+++ utils/checkAuthNum.ts
@@ -1,0 +1,7 @@
+export const checkAuthNum = (receivedAuthNum, authNum) => {
+  if (receivedAuthNum.toString() === authNum) {
+    return true;
+  } else {
+    return false;
+  }
+};
